MAIN Change: added preprocessorWrapper.py that should be useful in mass calling my closure compiler.
Closure Compiler is probably as done as it will get.
varPrepender.py has been modified for the preprocessorWrapper.py. It is
now easier to call from other python files and can append requires to
the front. Appending requires to the front ended up being useless, but
the functionallity is still there.